refactor (BookingColors): Move getEventColors() from mappers to utils
