+++ src/components/Calendar/Weeks/WeeksView.jsx
@@ -5,7 +5,7 @@ import classNames from "classnames";
 import "./Weeks.css";
 import fp from "lodash/fp";
 import { formatEvents } from "mappers/AppointmentMapper";
-import { getEventColors } from "mappers/RoomMapper";
+import { getEventColors } from "utils/BookingColors";
 
 const styles = theme => ({
   gridContainer: {
